changes to rotationCard - old api calling
